# Conflicts:
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/entity/AcquirerFee.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/service/AcquirerProvider.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/service/TransactionBuilder.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/service/TransactionSender.java

+++ services/authorization/src/test/java/com/processing/authorization/controller/AuthControllerTest.java
@@ -12,6 +12,8 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import static com.processing.authorization.constants.DeclineOutcome.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -48,7 +50,7 @@ class AuthControllerTest {
                 "123456",
                 "1234567890123456",
                 "000000",
-                5000L,
+                BigDecimal.valueOf(5000),
                 "810",
                 "2026-06-05T18:12:49.07",
                 "T0000001",
