+++ services/card-management/src/main/java/com/processing/cardmanagement/models/CardEntity.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.SQLRestriction;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -45,11 +46,11 @@ public class CardEntity {
     @Column(length = 3, nullable = false)
     private String currencyCode;
 
-    private long dailyLimit;
+    private BigDecimal dailyLimit;
 
-    private long monthlyLimit;
+    private BigDecimal monthlyLimit;
 
-    private long availableBalance;
+    private BigDecimal availableBalance;
 
     @Column(length = 10, nullable = false)
     private String issuerId;
