+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/repository/AcquirerFeeRepository.java
@@ -4,8 +4,10 @@ import com.processing.merchantacquirer.domain.entity.AcquirerFee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface AcquirerFeeRepository extends JpaRepository<AcquirerFee, Long> {
+public interface AcquirerFeeRepository extends JpaRepository<AcquirerFee, BigDecimal> {
     AcquirerFee findByTransmissionDateTimeAndStanAndTerminalIdAndAmountAndPan(
-            String transmissionDateTime, String stan, String terminalId, Long amount, String pan);
+            String transmissionDateTime, String stan, String terminalId, BigDecimal amount, String pan);
 }
